feat(TDP-5398): remove dataprep-dataset-client module
* use dataset.service.provider to select legacy or catalog dataset implementation
