--- dataprep-dataset-client/src/main/java/org/talend/dataprep/dataset/client/properties/DatasetProperties.java
@@ -1,37 +1,0 @@
-/*
- *  ============================================================================
- *
- *  Copyright (C) 2006-2018 Talend Inc. - www.talend.com
- *
- *  This source code is available under agreement available at
- *  https://github.com/Talend/data-prep/blob/master/LICENSE
- *
- *  You should have received a copy of the agreement
- *  along with this program; if not, write to Talend SA
- *  9 rue Pages 92150 Suresnes, France
- *
- *  ============================================================================
- */
-
-package org.talend.dataprep.dataset.client.properties;
-
-import java.net.URL;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties(prefix = "dataset.api")
-public class DatasetProperties {
-
-    public DatasetProperties() {}
-
-    private URL url;
-
-    public URL getUrl() {
-        return url;
-    }
-
-    public void setUrl(URL url) {
-        this.url = url;
-    }
-
-}
