+++ dataprep-dataset/pom.xml
@@ -29,11 +29,4 @@
         <stack.param.name>DataPrepDatasetFQIN</stack.param.name>
         <start-class>org.talend.dataprep.dataset.Application</start-class>
     </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.talend.dataprep</groupId>
-            <artifactId>dataprep-dataset-client</artifactId>
-        </dependency>
-    </dependencies>
 </project>
