+++ dataprep-dataset/src/main/java/org/talend/dataprep/dataset/Application.java
@@ -14,18 +14,15 @@ package org.talend.dataprep.dataset;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.Profile;
 import org.talend.dataprep.configuration.DataPrepComponentScanConfiguration;
-import org.talend.dataprep.dataset.client.properties.DatasetProperties;
 
 @SpringBootApplication
 @Configuration("org.talend.dataprep.dataset.Application")
 @Profile("standalone")
 @Import(DataPrepComponentScanConfiguration.class)
-@EnableConfigurationProperties({DatasetProperties.class })
 public class Application {
 
     public static void main(String[] args) {
