+++ dataprep-dataset/src/main/java/org/talend/dataprep/dataset/adapter/DataSetController.java
@@ -15,6 +15,8 @@
 
 package org.talend.dataprep.dataset.adapter;
 
+import java.util.List;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -22,10 +24,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.talend.dataprep.dataset.client.DatasetClient;
 import org.talend.dataprep.dataset.domain.Dataset;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/v1/datasets")
