+++ dataprep-dataset/src/main/java/org/talend/dataprep/dataset/adapter/DatasetSampleController.java
@@ -23,7 +23,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.talend.dataprep.dataset.client.DatasetClient;
 
 import com.fasterxml.jackson.databind.node.ObjectNode;
