+++ dataprep-dataset/src/main/java/org/talend/dataprep/dataset/configuration/DataSetConversions.java
@@ -13,19 +13,12 @@
 package org.talend.dataprep.dataset.configuration;
 
 import org.apache.commons.lang.StringUtils;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.talend.dataprep.api.dataset.DataSetMetadata;
 import org.talend.dataprep.api.share.Owner;
 import org.talend.dataprep.api.user.UserData;
 import org.talend.dataprep.conversions.BeanConversionService;
-import org.talend.dataprep.dataset.adapter.DataprepDatasetClient;
-import org.talend.dataprep.dataset.client.ProxyDatasetClient;
-import org.talend.dataprep.dataset.client.properties.DatasetProperties;
-import org.talend.dataprep.dataset.service.DataSetService;
 import org.talend.dataprep.dataset.service.UserDataSetMetadata;
 import org.talend.dataprep.processor.BeanConversionServiceWrapper;
 import org.talend.dataprep.security.Security;
@@ -63,18 +56,6 @@ public class DataSetConversions extends BeanConversionServiceWrapper {
                 }) //
                 .build());
         return conversionService;
-    }
-
-    @Bean
-    @ConditionalOnProperty("dataset.api.url")
-    public ProxyDatasetClient proxyDatasetClient(RestTemplateBuilder builder, DatasetProperties datasetProperties, Security security) {
-        return new ProxyDatasetClient(builder, datasetProperties, security);
-    }
-
-    @Bean
-    @ConditionalOnProperty(name = "dataset.api.url", havingValue = "disabled", matchIfMissing = true)
-    public DataprepDatasetClient dataprepDatasetClient(DataSetService dataSetService, BeanConversionService beanConversionService) {
-        return new DataprepDatasetClient(dataSetService, beanConversionService);
     }
 
 }
