+++ dataprep-dataset/src/main/resources/application.properties
@@ -35,8 +35,10 @@ dataset.local.file.size.limit=2000000000
 # Sets a limit on the number of data sets to be listed
 dataset.list.limit=10
 
+# Dataset provider (legacy or catalog)
+#dataset.service.provider=catalog
 # Talend Dataset Catalog url
-dataset.api.url=http://localhost:3000/api/v1
+#catalog.service.url=http://localhost:3000/api/v1
 
 # In memory dataset metadata store
 #dataset.metadata.store=in-memory
