+++ dataprep-dataset/src/test/java/org/talend/dataprep/dataset/adapter/MockDatasetServer.java
@@ -1,4 +1,21 @@
-package org.talend.dataprep.dataset.client;
+/*
+ *  ============================================================================
+ *
+ *  Copyright (C) 2006-2018 Talend Inc. - www.talend.com
+ *
+ *  This source code is available under agreement available at
+ *  https://github.com/Talend/data-prep/blob/master/LICENSE
+ *
+ *  You should have received a copy of the agreement
+ *  along with this program; if not, write to Talend SA
+ *  9 rue Pages 92150 Suresnes, France
+ *
+ *  ============================================================================
+ */
+
+package org.talend.dataprep.dataset.adapter;
+
+import java.io.IOException;
 
 import org.apache.commons.io.IOUtils;
 import org.springframework.http.HttpHeaders;
@@ -6,11 +23,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.io.IOException;
-
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.springframework.web.bind.annotation.RequestMethod.GET;
-import static org.talend.dataprep.dataset.client.ProxyDatasetClientTest.AUTHENTICATION_TOKEN;
+import static org.talend.dataprep.dataset.adapter.ProxyDatasetClientTest.AUTHENTICATION_TOKEN;
 
 @RestController
 @RequestMapping(value = "/api/v1", headers = HttpHeaders.AUTHORIZATION + "=" + AUTHENTICATION_TOKEN)
