+++ dataprep-dataset/src/test/java/org/talend/dataprep/dataset/adapter/ProxyDatasetClientTest.java
@@ -1,4 +1,22 @@
-package org.talend.dataprep.dataset.client;
+/*
+ *  ============================================================================
+ *
+ *  Copyright (C) 2006-2018 Talend Inc. - www.talend.com
+ *
+ *  This source code is available under agreement available at
+ *  https://github.com/Talend/data-prep/blob/master/LICENSE
+ *
+ *  You should have received a copy of the agreement
+ *  along with this program; if not, write to Talend SA
+ *  9 rue Pages 92150 Suresnes, France
+ *
+ *  ============================================================================
+ */
+
+package org.talend.dataprep.dataset.adapter;
+
+import java.net.MalformedURLException;
+import java.net.URL;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -6,11 +24,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.talend.dataprep.dataset.domain.Dataset;
-import org.talend.dataprep.dataset.client.properties.DatasetProperties;
 import org.talend.dataprep.security.Security;
-
-import java.net.MalformedURLException;
-import java.net.URL;
 
 import static org.junit.Assert.assertNotNull;
 import static org.mockito.Mockito.when;
@@ -29,9 +43,8 @@ public class ProxyDatasetClientTest extends TestParent {
 
     @Before
     public void updateClientUrl() throws MalformedURLException {
-        DatasetProperties datasetProperties = new DatasetProperties();
-        datasetProperties.setUrl(new URL("http://localhost:" + localServerPort + "/api/v1"));
-        datasetClient = new ProxyDatasetClient(restTemplateBuilder, datasetProperties, security);
+        URL url = new URL("http://localhost:" + localServerPort + "/api/v1");
+        datasetClient = new ProxyDatasetClient(restTemplateBuilder, url, security);
         when(security.getAuthenticationToken()).thenReturn(AUTHENTICATION_TOKEN);
     }
