+++ dataprep-dataset/src/test/java/org/talend/dataprep/dataset/adapter/TestParent.java
@@ -1,4 +1,19 @@
-package org.talend.dataprep.dataset.client;
+/*
+ *  ============================================================================
+ *
+ *  Copyright (C) 2006-2018 Talend Inc. - www.talend.com
+ *
+ *  This source code is available under agreement available at
+ *  https://github.com/Talend/data-prep/blob/master/LICENSE
+ *
+ *  You should have received a copy of the agreement
+ *  along with this program; if not, write to Talend SA
+ *  9 rue Pages 92150 Suresnes, France
+ *
+ *  ============================================================================
+ */
+
+package org.talend.dataprep.dataset.adapter;
 
 import org.junit.runner.RunWith;
 import org.springframework.boot.SpringBootConfiguration;
