+++ dataprep-dataset/src/test/resources/org/talend/dataprep/dataset/adapter/api_v1_datasets_d30e7e14-375a-49c7-87eb-62e1b614500d_withUiSpec_advanced.json
